Improved integer literal parsing.
Literals for smaller sized integer types are checked for bounds.

Unsigned literals are parsed correctly for their full range.

Added MOVABD instruction to load 64-bit immediate operands into
a register.

Allowed BasicValue to store a long, and generate code to load it into a
register if not fitting in a 32-bit signed integer for regular immediate
operands.

Allow underscores be used in integer literals.

+++ global_functions.cpp
@@ -61,7 +61,7 @@ Value *make_declaration_value(std::string name, TypeSpec *context) {
 }
 
 
-Value *make_basic_value(TypeSpec ts, int number) {
+Value *make_basic_value(TypeSpec ts, long number) {
     return new BasicValue(ts, number);
 }
